first source code format with new clang-tidy check
clang-tidy is a deep checker from the clang toolkit
It can mostly be used for code checking but also has an option to auto-
apply fixes for certain checks. These fixable checks are mostly concerned
with the c++-11 standard and are named modernize-*
 --> see list in codeStyleCheck.php

Currently applied are the checks
- modernize-use-nullptr -> changes all NULL-y values to a nullptr
- modernize-use-override -> marks functions as override, which override
base class functions

The clang-tidy checking takes some time (around 1 minute on my i7 cpu),
so the check is only optionallly available with the "checkTidy" flag.

To be discussed: which other checks should be enabled in the future.