fixed issue of pressing "M"-button causing alertdialog loop
this fixes issue #512
The button press was recognized as a long press, since the alertdialog
captures the focus and the button release did not correctly remove the
button from the actively-pressed-button-list.

Now the single-firing buttons do not get added to this list, which fixes
the issue.

Also fixed the handling of previous / next keyframe logic, which did not
modulo correct around the beginning and the end of the list.