At runtime replace $SHARED_DIR in paths (via FilePathHelper)
This is done before checking if the file exists, if you want a relative
path to a directory called "$SHARED_DIR", use "./$SHARED_DIR". The
reasoning for this is that this way, examples always appear as they are
intended to, even if there is a directory called "$SHARED_DIR" where
Mandelbulber was launched.

The following files have paths in them that I didn't know how to patch:
- Graeme McLaren  collection - license Creative Commons  (CC-BY 4.0)/mandelbulb_eye 001.fract
- Robert Pancoast collection - license Creative Commons (CC-BY 4.0)/box-deuce-deuce.fract
- Robert Pancoast collection - license Creative Commons (CC-BY 4.0)/menger-fatbackstrokin.fract
- mandelbox_menger_with_textures.fract
- monte carlo global illumination.fract

(From a grep -rl '[A-Z]:\\\|/home\|/usr' deploy/share/mandelbulber2/,
there might be more.)